% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normativeBRAIN-package.R
\docType{package}
\name{normativeBRAIN-package}
\alias{normativeBRAIN}
\alias{normativeBRAIN-package}
\title{normativeBRAIN: Application for normative curves of brain volumes using Weights for Incomplete Longitudinal Data and Quantile Regression}
\description{
Normative curves of brain volumes using package weightQuant : (i.e Viviane Philipps) estimation of observation-specific weights for incomplete longitudinal data and bootstrap procedure for weighted quantile regressions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arianebercu/normativeBRAIN}
  \item Report bugs at \url{https://github.com/arianebercu/normativeBRAIN/issues}
}

}
\keyword{internal}
